moved all config check function to a separate file
