+++ back-end/src/sciler/config/configHandler.go
@@ -234,163 +234,6 @@ func appendWhenUniqueComp(comps []*Component, comp *Component) []*Component {
 	return append(comps, comp)
 }
 
-// checkConfig is a method that will return an error
-// if the constraints value type is not equal to the device input type specified,
-// the actions type is not equal to the device output type,
-// or some other not allowed json configuration
-func checkConfig(config WorkingConfig) []string {
-	errList := make([]string, 0)
-	for _, puzzle := range config.Puzzles {
-		for _, rule := range puzzle.Event.Rules {
-			if err := rule.Conditions.checkConstraints(config, rule.ID); err != nil {
-				errList = append(errList, err...)
-			}
-			errList = append(errList, checkActions(rule.Actions, config)...)
-		}
-	}
-
-	for _, generalEvent := range config.GeneralEvents {
-		for _, rule := range generalEvent.Rules {
-			if err := rule.Conditions.checkConstraints(config, rule.ID); err != nil {
-				errList = append(errList, err...)
-			}
-			errList = append(errList, checkActions(rule.Actions, config)...)
-		}
-	}
-
-	for _, rule := range config.ButtonEvents {
-		if err := rule.Conditions.checkConstraints(config, rule.ID); err != nil {
-			errList = append(errList, err...)
-		}
-		errList = append(errList, checkActions(rule.Actions, config)...)
-	}
-
-	// TODO check uniqueness of all device_id, timer_id and rule_id
-	return errList
-}
-
-// checkAction is a method that will return an error
-// if the action's value types and instructions are not equal to the unit's output specifications
-func checkActions(actions []Action, config WorkingConfig) []string {
-	errorList := make([]string, 0)
-	for _, action := range actions {
-		switch action.Type {
-		case "device":
-			errorList = append(errorList, checkActionDevice(action, config)...)
-		case "timer":
-			errorList = append(errorList, checkActionTimer(action, config)...)
-		case "label":
-			errorList = append(errorList, checkActionLabel(action, config)...)
-		default:
-			errorList = append(errorList,
-				fmt.Sprintf("only device, timer and label are accepted as type for an action, however type was specified as: %s", action.Type))
-		}
-	}
-	return errorList
-}
-
-// checkActionTimer is a method that checks the config in use for mistakes in the action of a timer
-// if the config does not follow the manual, a non-empty list of mistakes is returned
-func checkActionTimer(action Action, config WorkingConfig) []string {
-	errorList := make([]string, 0)
-	if _, ok := config.Timers[action.TypeID]; ok { // checks if timer can be found in the map, if so, it is stored in variable device
-		for _, actionMessage := range action.Message {
-			switch actionMessage.Instruction {
-			case "add", "subtract":
-				{
-					valueType := reflect.TypeOf(actionMessage.Value).Kind()
-					if valueType != reflect.String {
-						errorList = append(errorList,
-							fmt.Sprintf("input type string expected but %s found as type of value %v",
-								valueType.String(), actionMessage.Value))
-					}
-					break
-				}
-			case "start", "pause", "stop", "done":
-				break
-			default:
-				errorList = append(errorList, fmt.Sprintf("instruction %s is not defined for a timer", actionMessage.Instruction))
-			}
-		}
-	} else {
-		errorList = append(errorList, fmt.Sprintf("timer with id %s not found in map", action.TypeID))
-	}
-	return errorList
-}
-
-// checkActionDevice is a method that checks the current config for mistakes in the action of a device
-// if the config does not follow the manual, a non-empty list of mistakes is returned
-func checkActionDevice(action Action, config WorkingConfig) []string {
-	errorList := make([]string, 0)
-	if device, ok := config.Devices[action.TypeID]; ok { // checks if device can be found in the map, if so, it is stored in variable device
-		for _, actionMessage := range action.Message {
-			if outputObject, ok := device.Output[actionMessage.ComponentID]; ok {
-				if instructionType, ok := outputObject.Instructions[actionMessage.Instruction]; ok {
-					if err := checkActionInstructionType(reflect.TypeOf(actionMessage.Value).Kind(), instructionType, actionMessage.Value); err != nil {
-						errorList = append(errorList, err.Error())
-					}
-				} else {
-					errorList = append(errorList, fmt.Sprintf("instruction %s not found in map", actionMessage.Instruction))
-				}
-			} else {
-				errorList = append(errorList, fmt.Sprintf("component with id %s not found in map", actionMessage.ComponentID))
-			}
-		}
-	} else {
-		errorList = append(errorList, fmt.Sprintf("device with id %s not found in map", action.TypeID))
-	}
-	return errorList
-}
-
-// checkActionInstructionType checks if the type op the value of an instruction is the same as the type the instruction requires according to the config
-func checkActionInstructionType(valueType reflect.Kind, instructionType string, value interface{}) error {
-	switch instructionType {
-	case "string":
-		if valueType != reflect.String {
-			return fmt.Errorf("instruction type string expected but %s found as type of value %v",
-				valueType.String(), value)
-		}
-	case "boolean":
-		if valueType != reflect.Bool {
-			return fmt.Errorf("instruction type boolean expected but %s found as type of value %v",
-				valueType.String(), value)
-		}
-	case "numeric":
-		if valueType != reflect.Int && valueType != reflect.Float64 {
-			return fmt.Errorf("instruction type numeric expected but %s found as type of value %v",
-				valueType.String(), value)
-		}
-	case "array":
-		if valueType != reflect.Slice {
-			return fmt.Errorf("instruction type array/slice expected but %s found as type of value %v",
-				valueType.String(), value)
-		}
-	default:
-		return fmt.Errorf("custom types like: %s, are not yet implemented", instructionType)
-	}
-	return nil
-}
-
-// checkActionLabel checks if there is a label with this ID,
-// and checks if all components under this label have the correct instructions with a call to checkActionDevice
-// if the config does not follow the manual, a non-empty list of mistakes is returned
-func checkActionLabel(action Action, config WorkingConfig) []string {
-	errorList := make([]string, 0)
-	if _, ok := config.LabelMap[action.TypeID]; ok { // checks if label can be found in the map, if so, it is stored in variable device
-		for _, instruction := range action.Message {
-			for _, comp := range config.LabelMap[action.TypeID] {
-				instruction.ComponentID = comp.ID
-				errorList = append(errorList,
-					checkActionDevice(Action{TypeID: comp.Device.ID, Message: []ComponentInstruction{instruction}}, config)...)
-			}
-		}
-	} else {
-		errorList = append(errorList,
-			fmt.Sprintf("label with id %s not found in map", action.TypeID))
-	}
-	return errorList
-}
-
 // generatePuzzles transforms readPuzzles to puzzles
 // it generates events and copies the rest
 // if the config does not follow the manual, a non-empty list of mistakes is returned
